@@ -11,16 +11,21 @@
     },
     "devDependencies": {
         "@popperjs/core": "^2.10.2",
-        "axios": "^0.21",
-        "bootstrap": "^5.1.3",
+        "axios": "^0.21.4",
+        "bootstrap": "^5.2.3",
         "chokidar": "^3.5.3",
         "laravel-mix": "^6.0.6",
-        "lodash": "^4.17.19",
+        "lodash": "^4.17.21",
         "postcss": "^8.1.14",
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
-        "vue": "^2.6.12",
-        "vue-template-compiler": "^2.6.12"
+        "ts-loader": "^9.4.2",
+        "typescript": "^4.9.3",
+        "vue": "^3.2.45",
+        "vue-axios": "^3.5.2",
+        "vue-loader": "^17.0.1",
+        "vue-template-compiler": "^2.6.12",
+        "vuex": "^4.1.0"
     }
 }
