@@ -4,29 +4,19 @@
  * building robust, powerful web applications using Vue and Laravel.
  */
 
-require('./bootstrap');
+require('./bootstrap')
+import axios from 'axios'
+import VueAxios from 'vue-axios'
+import { createApp } from 'vue'
+import statIndex from './pages/statistic/Index'
+import AccessToken from "./components/AccessToken";
 
-window.Vue = require('vue').default;
+import store from './store'
+const app = createApp({})
+app.component('stat-index', statIndex)
+app.component('access-token', AccessToken)
+axios.defaults.withCredentials = true;
+app.use(store);
+app.use(VueAxios, axios)
+app.mount('#app')
 
-/**
- * The following block of code may be used to automatically register your
- * Vue components. It will recursively scan this directory for the Vue
- * components and automatically register them with their "basename".
- *
- * Eg. ./components/ExampleComponent.vue -> <example-component></example-component>
- */
-
-// const files = require.context('./', true, /\.vue$/i)
-// files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default))
-
-Vue.component('example-component', require('./components/ExampleComponent.vue').default);
-
-/**
- * Next, we will create a fresh Vue application instance and attach it to
- * the page. Then, you may begin adding components to this application
- * or customize the JavaScript scaffolding to fit your unique needs.
- */
-
-const app = new Vue({
-    el: '#app',
-});
